+++ src/cache/get_all_clusters.rs
@@ -1,0 +1,12 @@
+use super::Cache;
+
+pub trait GetAllClusters {
+    fn get_all_clusters(&self) -> Vec<String>;
+}
+
+impl GetAllClusters for Cache {
+    fn get_all_clusters(&self) -> Vec<String> {
+        let store = self.store.lock().unwrap();
+        store.keys().cloned().collect()
+    }
+}
