+++ src/cache/mod.rs
@@ -1,3 +1,13 @@
 pub mod cache;
+pub mod clear_all;
+pub mod clear_cluster;
+pub mod clear_expired;
+pub mod decr;
+pub mod delete;
+pub mod get;
+pub mod get_all_clusters;
+pub mod get_cluster_keys;
 pub mod incr;
+pub mod set;
+pub mod set_cluster;
 pub use cache::Cache;
